Fixed build problems
This commit fixes a few issues with the build. First, it adds an explicit
Breeze dependency, because otherwise there's an inconsistency where the
Breeze version that this project gets compiled with is older than the
one that Spark has.

Second, it fixes some Hadoop compatibility issues. This project needs
a Hadoop dependency explicitly because otherwise Spark pulls in the
wrong Hadoop version. We use 2.0.0-mr1-cdh4.2.0, which is *not* the
version monotasks uses (monotasks uses 2.0.0-cdh4.2.0), because otherwise
there are some suspicious build issues where Maven can't find some of the
YARN libraries. The mr1 version is API-compatible with the one monotasks
uses, so this makes compilation work, and then Hadoop is marked is provided,
so the bad verison isn't included in the assembly JAR.